requirements.txt
python-dotenv==0.10.5
requests==2.22.0
pyperclip==1.7.0
